+++ src/django/api/serializers.py
@@ -422,7 +422,9 @@ class FacilitySerializer(GeoFeatureModelSerializer):
     class Meta:
         model = Facility
         fields = ('id', 'name', 'address', 'country_code', 'location',
-                  'oar_id', 'country_name', 'contributors')
+                  'oar_id', 'country_name', 'contributors',
+                  'ppe_product_types', 'ppe_contact_phone',
+                  'ppe_contact_email', 'ppe_website')
         geo_field = 'location'
 
     # Added to ensure including the OAR ID in the geojson properties map